Move HealthSummary and HealthIcon to pages
