+++ assets/js/components/SapSystemDetails/tableConfigs.jsx
@@ -5,7 +5,7 @@ import HostLink from '@components/HostLink';
 import ProviderLabel from '@components/ProviderLabel';
 import CleanUpButton from '@components/CleanUpButton';
 import Tooltip from '@common/Tooltip';
-import HealthIcon from '@components/Health';
+import HealthIcon from '@pages/HealthIcon';
 
 import Features from './Features';
 import InstanceStatus from './InstanceStatus';
