+++ assets/js/components/SapSystemsOverview/SapSystemsOverview.jsx
@@ -6,13 +6,13 @@ import { filter } from 'lodash';
 import { getEnsaVersionLabel } from '@lib/model/sapSystems';
 
 import PageHeader from '@common/PageHeader';
-import HealthIcon from '@components/Health';
+import HealthIcon from '@pages/HealthIcon';
 import Table from '@common/Table';
 import SAPSystemItemOverview from '@components/SapSystemsOverview/SapSystemItemOverview';
 import Tags from '@common/Tags';
-import HealthSummary from '@components/HealthSummary/HealthSummary';
+import HealthSummary from '@pages/HealthSummary';
 import DeregistrationModal from '@components/DeregistrationModal';
-import { getCounters } from '@components/HealthSummary/summarySelection';
+import { getCounters } from '@pages/HealthSummary/summarySelection';
 
 function SapSystemsOverview({
   sapSystems,
