+++ assets/js/pages/HealthIcon/HealthIcon.stories.jsx
