+++ assets/js/pages/HealthIcon/index.js
