+++ assets/js/pages/HealthSummary/HomeHealthSummary.jsx
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import Table from '@common/Table';
-import HealthIcon from '@components/Health/HealthIcon';
+import HealthIcon from '@pages/HealthIcon';
 import { Link } from 'react-router-dom';
 import PageHeader from '@common/PageHeader';
-import HealthSummary from '@components/HealthSummary';
+import HealthSummary from '@pages/HealthSummary';
 import useQueryStringValues from '@hooks/useQueryStringValues';
 import { getCounters, isMostRelevantPrio } from './summarySelection';
